리뷰하기 기능 구현

+++ src/main/java/com/example/demo/siteuser/dto/MyInfoDto.java
@@ -17,7 +17,7 @@ public class MyInfoDto {
     private String nickname;
     private String email;
     private String phoneNumber;
-    private double mannerScore;
+    private int mannerScore;
     private GenderType gender;
     private Ntrp ntrp;
     private String address;
